lazy.ts in types/create
