+++ packages/types/src/create/index.ts
@@ -5,5 +5,6 @@ export * from './createClass';
 export * from './createType';
 export * from './encodeTypes';
 export * from './getTypeDef';
+export * from './lazy';
 export * from './registry';
 export * from './typeSplit';
